grep_topwords.py: merge the dic files and show the top most words which occur in those dics.
Usage: change the dic filepaths in the file, and run it directly.

To know which words are proper to recognize one author's writing
pattern, this program will calculate the occurrence times of each
word in the given dic files, and list out the top most N words.

Note that the user must manually check the word list to pick proper
words. It is very possible to get bad performance without human
intellegence.