run_test.py: the training process.
Usage: python3 [this]

To offer the training data and testing data, the user can change
the contents of train_data and to_test, respectively. The variable
black_list is a list of the items to be filtered out in train_data,
which makes it more convenient to compare the performance of different
training datasets.

To change the model parameters, just modify the input arguments of
train_set. The second argument is nu. The smaller nu is, the looser
model edge is. However, currently due to some bug of libsvm, DO NOT
USE ANY VALUE OTHER THAN {0.1, 0.2, ..., 1.0}. The third parameter,
gamma, can control the edge as well. The property of gamma is similar
to nu.

By my experiments, if possible, keeping nu to be 0.5 and tuning only
gamma parameter will be better. On the other hand, due to the bug
mentioned above, gamma is more 'tunable'.