make_3d_graph.py: use pyplot to get a 3d graph.
To visually show the difference between datasets,
we can use PAC to reduce the dimension to 3, and
then use pyplot to draw a scatter plot.

Since there are so many technique issues in using
this file, including the installation of numpy,
scipy and matplotlib, this file needs more
background skills to use.

Currently it is used to show that bm (白馬嘯西風)
and lchen (連城訣) are very different from other
金庸's work. To let the user know where the
classification bound is, predicted positive
data and predicted negtive data are drawn in
different colors.