filepaths.py: change the name of the dic files.
To make it consistent, here all dic files are
changed to the same format.

If a book has only one chapter, then the only
dic file will be named 'dic'. Otherwise, the
chapters of the book will be named dic_[number],
where the number is the order of the chapters.
The length of leading zeros depends on the number
of chapters in the book.