freq_comp.py: a word frequency analysis tool.
If we use the model and find that some chapter/work is
abnormal, we may want to take a deeper look at the
word frequency of the text. This file is just the tool.

In this file, set1 is a list, which contains the works
we pick to be 'normal work set'. On the other hand, set2,
is the work/chapter we want to take a look.

In the current version, we show the frequency difference
by calculating relative frequency of words. To make sure
that each word occurs in set1, now we only focus on words
in feature_list.

By the way, since this file is in ./tools and it needs to
access filepaths.py, this commit also does some changes
on filepaths.py. It is expected to be compatible with other
files which also import filepaths.py.

TEST=PASSED
Both run_test.py and freq_comp.py work well by executing
them in their current directory.